Added outline interactions
Any static mesh component with collision presets set to 'Custom', object type to 'Interactable' and with a component tag element 'Outline' would have a highlighted outline when character approaches. For now I use it for object that needs lmb clicking so it prints 'click to select', could optimise my input key to make it compatible with others'.